Add config in main to run gtag and measure on the same page
