+++ demo/public/index.html
@@ -1,10 +1,23 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <script async src="./measure.js"></script>
+    <!-- Include gtag... data layer commands conflict with measure though -->
+    <script async src="https://www.googletagmanager.com/gtag/js?id=%REACT_APP_GA_TRACKING_ID%"></script>
     <script>
-      window.dataLayer = window.dataLayer || [];
-      function measure(){dataLayer.push(arguments);}
+    window.dataLayer = window.dataLayer || [];
+    function gtag(){dataLayer.push(arguments);}
+    gtag('js', new Date());
+    </script>
+    <!-- Include measurement library. Normally we can't run this and gtag on
+         the same site, because both use the same dataLayer object and send
+         conflicting events. As a workaround, we do not run async and include
+         a measure-setup compiled file that does not run commands on load,
+         then call it in the setup script. -->
+    <script src="./measure-no-config.js"></script>
+    <script>
+      measureDataLayer = [];
+      window.setupMeasure(measureDataLayer);
+      function measure(){measureDataLayer.push(arguments);}
       measure("config", "googleAnalytics", {
             "measurement_id": "%REACT_APP_GA_TRACKING_ID%",
             "api_secret": "%REACT_APP_GA_API_SECRET%"
