--- demo/public/measure.js
@@ -1,19 +1,0 @@
-(function(){/*
-
- Copyright The Closure Library Authors.
- SPDX-License-Identifier: Apache-2.0
-*/
-'use strict';var f=this||self;/*
- jQuery v1.9.1 (c) 2005, 2012
- jQuery Foundation, Inc. jquery.org/license.
-*/
-var g=/\[object (Boolean|Number|String|Function|Array|Date|RegExp|Arguments)\]/;function k(a){return null==a?String(a):(a=g.exec(Object.prototype.toString.call(Object(a))))?a[1].toLowerCase():"object"}function l(a,b){return Object.prototype.hasOwnProperty.call(Object(a),b)}function n(a){if(!a||"object"!=k(a)||a.nodeType||a==a.window)return!1;try{if(a.constructor&&!l(a,"constructor")&&!l(a.constructor.prototype,"isPrototypeOf"))return!1}catch(c){return!1}for(var b in a);return void 0===b||l(a,b)};function q(a,b){var c={},d=c;a=a.split(".");for(var e=0;e<a.length-1;e++)d=d[a[e]]={};d[a[a.length-1]]=b;return c}function t(a,b){var c=!a._clear,d;for(d in a)if(l(a,d)){var e=a[d];"array"===k(e)&&c?("array"===k(b[d])||(b[d]=[]),t(e,b[d])):n(e)&&c?(n(b[d])||(b[d]={}),t(e,b[d])):b[d]=e}delete b._clear};/*
- Copyright 2012 Google Inc. All rights reserved. */
-function u(a,b,c){b=void 0===b?{}:b;"function"===typeof b?b={listener:b,listenToPast:void 0===c?!1:c,processNow:!0,commandProcessors:{}}:b={listener:b.listener||function(){},listenToPast:b.listenToPast||!1,processNow:void 0===b.processNow?!0:b.processNow,commandProcessors:b.commandProcessors||{}};this.a=a;this.l=b.listener;this.j=b.listenToPast;this.g=!1;this.c={};this.f=[];this.b=b.commandProcessors;this.h=v(this);b.processNow&&this.process()}
-u.prototype.process=function(){this.registerProcessor("set",function(){var c={};1===arguments.length&&"object"===k(arguments[0])?c=arguments[0]:2===arguments.length&&"string"===k(arguments[0])&&(c=q(arguments[0],arguments[1]));return c});w(this,this.a,!this.j);var a=this.a.push,b=this;this.a.push=function(){var c=[].slice.call(arguments,0),d=a.apply(b.a,c);w(b,c);return d}};u.prototype.get=function(a){var b=this.c;a=a.split(".");for(var c=0;c<a.length;c++){if(void 0===b[a[c]])return;b=b[a[c]]}return b};
-u.prototype.flatten=function(){this.a.splice(0,this.a.length);this.a[0]={};t(this.c,this.a[0])};u.prototype.registerProcessor=function(a,b){a in this.b||(this.b[a]=[]);this.b[a].push(b)};
-function w(a,b,c){c=void 0===c?!1:c;for(a.f.push.apply(a.f,b);!1===a.g&&0<a.f.length;){b=a.f.shift();if("array"===k(b))a:{var d=a.c;k(b[0]);for(var e=b[0].split("."),m=e.pop(),p=b.slice(1),h=0;h<e.length;h++){if(void 0===d[e[h]])break a;d=d[e[h]]}try{d[m].apply(d,p)}catch(F){}}else if("arguments"===k(b)){e=a;m=[];p=b[0];if(e.b[p])for(d=e.b[p].length,h=0;h<d;h++)m.push(e.b[p][h].apply(e.h,[].slice.call(b,1)));a.f.push.apply(a.f,m)}else if("function"==typeof b)try{b.call(a.h)}catch(F){}else if(n(b))for(var r in b)t(q(r,
-b[r]),a.c);else continue;c||(a.g=!0,a.l(a.c,b),a.g=!1)}}u.prototype.registerProcessor=u.prototype.registerProcessor;u.prototype.flatten=u.prototype.flatten;u.prototype.get=u.prototype.get;u.prototype.process=u.prototype.process;window.DataLayerHelper=u;function v(a){return{set:function(b,c){t(q(b,c),a.c)},get:function(b){return a.get(b)}}};function x(a){console.error(a)};function y(a){var b=void 0===a?{}:a,c=b.api_secret,d=b.measurement_id,e=b.measurement_url;a=b.client_id_expires;b=b.automatic_params;void 0!==c&&"string"!==typeof c&&(x("Provided API Secret is not a valid string."),c=void 0);void 0!==d&&"string"!==typeof d&&(x("Provided Measurement ID is not a valid string."),d=void 0);"string"!==typeof e&&(void 0!==e&&x("Provided Measurement URL is not a valid string. Using default URL instead."),void 0!==d&&void 0!==c||x("Sending events to Google Analytics without measurement ID or API secret will lead to events being dropped."));
-isNaN(Number(a))?(void 0!==a&&x("Provided Client ID Expires is not a valid number. Using default value instead."),a=63072E3):a=Number(a);"[object Array]"!==Object.prototype.toString.call(b)&&(void 0!==b&&x("Provided Automatic Parameters are not contained within an array Discarding provided value."),b=[]);this.a={page_path:!0,page_location:!0,page_title:!0,user_id:!0,client_id:!0};for(c=0;c<b.length;++c)this.a[b[c]]=!0;this.b=a}y.prototype.i=function(){return"googleAnalytics"};
-y.prototype.processEvent=function(){};y.prototype.persistTime=function(a){return"client_id"===a?this.b:-1};y.prototype.persistTime=y.prototype.persistTime;y.prototype.processEvent=y.prototype.processEvent;y.prototype.getName=y.prototype.i;var z={},A={googleAnalytics:y};function B(a,b){"string"===typeof a&&(a in z||x("No storage interface with name "+a+" found."),a=z[a]);try{return new a(b)}catch(c){return x("Could not construct a new instance of the class "+(a+" with parameter")),x(b),x(c),null}}function C(a,b){"string"===typeof a&&(a in A||x("No event processor with name "+a+" found."),a=A[a]);try{return new a(b)}catch(c){return x("Could not construct a new instance of the class "+(a+" with parameter")),x(b),x(c),null}}
-function D(a,b,c,d,e){var m=C(b,c),p=B(d,e);m&&p&&a.registerProcessor("event",function(h,r){r||(r={});m.processEvent(p,this,h,r)})};function E(a){var b=new u(a,{processNow:!1});b.registerProcessor("config",function(c,d,e,m){D(b,c,d,e,m)});b.process()}var G=["setupMeasure"],H=f;G[0]in H||"undefined"==typeof H.execScript||H.execScript("var "+G[0]);for(var I;G.length&&(I=G.shift());)G.length||void 0===E?H[I]&&H[I]!==Object.prototype[I]?H=H[I]:H=H[I]={}:H[I]=E;window.dataLayer=window.dataLayer||[];E(window.dataLayer);})();
